RDM-12742 Fix CaseUserControllerIT
