+++ src/test/java/uk/gov/hmcts/ccd/v2/external/controller/BaseCaseAssignedUserRolesControllerIT.java
@@ -265,26 +265,4 @@ class BaseCaseAssignedUserRolesControllerIT extends WireMockBaseTest {
             return 0L;
         }
     }
-
-    protected void stubUserInfo(String userId) {
-        stubFor(WireMock.get(urlMatching("/o/userinfo"))
-            .willReturn(okJson("{"
-                + "      \"uid\": \"" + userId + "\","
-                + "      \"sub\": \"Cloud.Strife@test.com\","
-                + "      \"roles\": [ \"caseworker\", \"caseworker-test\", \"caseworker-PROBATE-public\","
-                + " \"caseworker-PROBATE\", \"caseworker-DIVORCE\", \"caseworker-SSCS\" ]"
-                + "    }").withStatus(200)));
-    }
-
-    protected void stubIdamRolesForUser(String userId) {
-        stubFor(WireMock.get(urlMatching("/api/v1/users/" + userId))
-            .willReturn(okJson("{"
-                + "      \"id\": \" " + userId + "\","
-                + "      \"email\": \"Cloud.Strife@test.com\","
-                + "      \"forename\": \"Cloud\","
-                + "      \"surname\": \"Strife\","
-                + "      \"roles\": [ \"caseworker\", \"caseworker-test\", \"caseworker-PROBATE-public\","
-                + " \"caseworker-PROBATE\", \"caseworker-DIVORCE\", \"caseworker-SSCS\" ]"
-                + "    }").withStatus(200)));
-    }
 }
